fix: Avoid sending raw path for invalid routes as operation names (#235)
We were setting up req.path as operation name when creating the middleware
this is undefined if you set the tracing middleware as the first one, however
it tends to be defined if you set that after URL parsing.

The placeholder was then replaced by the route on finish, however, if there
was not a matching route (e.g. a request to login.auth0.com/12345) the original
path was preserved as operation name. This is problematic because we end up
with 100s of useless operations and the UI becomes clumsy and useless, full
of invalid operations example: https://app.lightstep.com/auth0-prod/service-directory/auth0-api1/operations